+++ R/get_PI_dates.R
@@ -64,7 +64,7 @@ get_PI_dates <- function(mod,
                             value.name = "hour")
 
   # convert indx to date_time
-  dat_l[,hour := mod$start_time + (hour*3600)]
-  dat_l
-  return(print(dat_l))
+  dat_l[,hour := mod$start_time + (hour*3600)][]
+
+  return(dat_l)
 }