+++ NEWS.md
@@ -1,4 +1,5 @@
 # viticolaR 0.0.4
+ * Add summarising method function for `m_viticolaR` objects
  * Add ggplot geom layers for plotting  
  * New feature - plot_weather()  
  * Example fixes in 'Start' and end 'times' which effect timing